[INLONG-9867][Manager] Fix error
